pyproject.toml
[tool.poetry]
name = "drfproject-onlineschool"
version = "0.1.0"
description = ""
authors = ["Sumerian <shemerov.s@yandex.ru>"]
readme = "README.md"
packages = [{include = "drfproject_onlineschool"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
black = "^24.8.0"
coverage = "^7.6.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
requests = "^2.32.3"
stripe = "^10.12.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
